refactor: rename system state
