+++ src/extension/features/system/slice.ts
@@ -5,7 +5,7 @@ import { StoreNameEnum } from '@extension/enums';
 
 // types
 import type { ILogger } from '@common/types';
-import type { IConfirmModal, IScanQRCodeModal, ISystemState } from './types';
+import type { IConfirmModal, IScanQRCodeModal, IState } from './types';
 
 // utils
 import { getInitialState } from './utils';
@@ -15,27 +15,24 @@ const slice = createSlice({
   name: StoreNameEnum.System,
   reducers: {
     setConfirmModal: (
-      state: Draft<ISystemState>,
+      state: Draft<IState>,
       action: PayloadAction<IConfirmModal | null>
     ) => {
       state.confirmModal = action.payload;
     },
-    setLogger: (state: Draft<ISystemState>, action: PayloadAction<ILogger>) => {
+    setLogger: (state: Draft<IState>, action: PayloadAction<ILogger>) => {
       state.logger = action.payload;
     },
-    setOnline: (state: Draft<ISystemState>, action: PayloadAction<boolean>) => {
+    setOnline: (state: Draft<IState>, action: PayloadAction<boolean>) => {
       state.online = action.payload;
     },
     setScanQRCodeModal: (
-      state: Draft<ISystemState>,
+      state: Draft<IState>,
       action: PayloadAction<IScanQRCodeModal | null>
     ) => {
       state.scanQRCodeModal = action.payload;
     },
-    setSideBar: (
-      state: Draft<ISystemState>,
-      action: PayloadAction<boolean>
-    ) => {
+    setSideBar: (state: Draft<IState>, action: PayloadAction<boolean>) => {
       state.sidebar = action.payload;
     },
   },
