+++ src/extension/features/system/types/index.ts
@@ -1,3 +1,3 @@
 export type { default as IConfirmModal } from './IConfirmModal';
 export type { default as IScanQRCodeModal } from './IScanQRCodeModal';
-export type { default as ISystemState } from './ISystemState';
+export type { default as IState } from './IState';
