+++ src/extension/types/states/IBaseRootState.ts
@@ -1,6 +1,6 @@
 // features
 import type { IARC0200AssetsState } from '@extension/features/arc0200-assets';
-import type { ISystemState } from '@extension/features/system';
+import type { IState as ISystemState } from '@extension/features/system';
 
 interface IBaseRootState {
   arc0200Assets: IARC0200AssetsState;
