+++ src/extension/types/states/IRegistrationRootState.ts
@@ -3,7 +3,7 @@ import type { IState as INetworksState } from '@extension/features/networks';
 import type { INotificationsState } from '@extension/features/notifications';
 import type { IState as RegistrationState } from '@extension/features/registration';
 import type { ISettingsState } from '@extension/features/settings';
-import type { ISystemState } from '@extension/features/system';
+import type { IState as ISystemState } from '@extension/features/system';
 
 // types
 import type IBaseRootState from './IBaseRootState';
